+++ tests/test_surround.lua
@@ -541,22 +541,22 @@ T['Add surrounding']['works with multibyte characters'] = function()
 end
 
 T['Add surrounding']['works on whole line'] = function()
-  -- Should ignore indent at left mark but not whitespace at right
-  -- Should work with both tabs and spaces
-  validate_edit({ ' \t aaa ', '' }, { 1, 0 }, { ' \t (aaa )', '' }, { 1, 4 }, type_keys, 'sa', '_', ')')
-  validate_edit({ ' \t aaa ', '' }, { 1, 0 }, { ' \t (aaa )', '' }, { 1, 4 }, type_keys, 'V', 'sa', ')')
+  -- Should ignore both indent (leading whitespace) at start line and trailing
+  -- whitespace and end line. Should work with both tabs and spaces.
+  validate_edit({ ' \t aaa\t ', '' }, { 1, 0 }, { ' \t (aaa)\t ', '' }, { 1, 4 }, type_keys, 'sa', '_', ')')
+  validate_edit({ ' \t aaa\t ', '' }, { 1, 0 }, { ' \t (aaa)\t ', '' }, { 1, 4 }, type_keys, 'V', 'sa', ')')
 end
 
 T['Add surrounding']['works on multiple lines'] = function()
   local f = function() type_keys('sa', 'ap', ')') end
   local f_vis = function() type_keys('Vap', 'sa', ')') end
 
-  -- Should ignore indent at left mark but not whitespace at right
-  -- Should work with both tabs and spaces
-  validate_edit({ ' \t aaa ', 'bbb', ' ccc' }, { 1, 0 }, { ' \t (aaa ', 'bbb', ' ccc)' }, { 1, 4 }, f)
-  validate_edit({ ' \t aaa ', 'bbb', ' ccc' }, { 1, 0 }, { ' \t (aaa ', 'bbb', ' ccc)' }, { 1, 4 }, f_vis)
-  validate_edit({ ' \t aaa ', ' ' }, { 1, 0 }, { ' \t (aaa ', ' )' }, { 1, 4 }, f)
-  validate_edit({ ' \t aaa ', ' ' }, { 1, 0 }, { ' \t (aaa ', ' )' }, { 1, 4 }, f_vis)
+  -- Should ignore both indent (leading whitespace) at start line and trailing
+  -- whitespace and end line. Should work with both tabs and spaces.
+  validate_edit({ ' \t aaa ', 'bbb', ' ccc\t ' }, { 1, 0 }, { ' \t (aaa ', 'bbb', ' ccc)\t ' }, { 1, 4 }, f)
+  validate_edit({ ' \t aaa ', 'bbb', ' ccc\t ' }, { 1, 0 }, { ' \t (aaa ', 'bbb', ' ccc)\t ' }, { 1, 4 }, f_vis)
+  validate_edit({ ' \t aaa ', '\t ' }, { 1, 0 }, { ' \t (aaa ', ')\t ' }, { 1, 4 }, f)
+  validate_edit({ ' \t aaa ', '\t ' }, { 1, 0 }, { ' \t (aaa ', ')\t ' }, { 1, 4 }, f_vis)
 end
 
 T['Add surrounding']['works with multiline output surroundings'] = function()